fix: wip rearrange samples; add CropOnControlPoints and migrate internal classes
