+++ samples/1-mobile-camera/config.json
