+++ samples/1-mobile-camera/omr_marker.jpg
