+++ samples/1-mobile-camera/template.json
