+++ samples/3-answer-key/using-image/config.json
