+++ samples/3-answer-key/using-image/evaluation.json
