+++ samples/3-answer-key/using-image/template.json
